requirements.txt
asgiref==3.8.1
Django==5.2
djangogrpcframework==0.2.1
djangorestframework==3.16.0
grpcio==1.71.0
grpcio-reflection==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
setuptools==80.3.0
sqlparse==0.5.3
